Update README.md
Gestational age is a method of studying the patterns of fetal growth and development over time based on gestational age at birth. Gestational age is the number of weeks that have passed since the first day of the last menstrual period. It is an important indicator of fetal maturity and health outcomes